requirements.txt
cvxopt==1.3.2
einops==0.8.1
matplotlib==3.10.3
numpy==2.2.5
opencv_python==4.10.0.84
Pillow==11.2.1
scipy==1.15.3
scikit-image
timm==1.0.15
torch
torchvision
tqdm==4.67.0
fast_pytorch_kmeans
psutil